Update Base Dependencies
Update
- Pillow
- tensorflow
- torch
- torchvision

Removed requirements.txts in specific pages subdirectories
because they should have their needed dependencies
installed from the base requirements.txt and it is a pain
to keep multiple copies up to date with each other.

Removed h5py, and Keras explicit dependencies because it
seems the most recent version of tensorflow doesn't have
the same problem? Tested the image classifier page and
it classified fine on newer versions of tensorflow.
If we don't do that, then tensorflow won't work with
python 3.9, since all versions of tensorflow that work
with 3.9 require h5py > 3.0